+++ conversores/ParaRomano.cpp
@@ -1,8 +1,9 @@
 #include "ParaRomano.hpp"
 
 #include <string>
+#include <cstdint>
 
-std::string ParaRomano::DeLong(unsigned long numToConvert){
+std::string ParaRomano::DeLong(uint32_t numToConvert){
 	std::string resultado = "";
 	for (size_t i = 0; i < std::size(ParaRomano::VALORES); i++) {
 		while (numToConvert >= ParaRomano::VALORES[i]) {
